Move APIViewTestCases to api.py
