+++ netbox/utilities/testing/__init__.py
@@ -1,2 +1,3 @@
-from .testcases import *
+from .api import *
 from .utils import *
+from .views import *
